拡張子変換処理をmainパッケージへ移動
